+++ src/main/java/com/ourmenu/backend/domain/user/api/UserController.java
@@ -1,12 +1,10 @@
 package com.ourmenu.backend.domain.user.api;
 
-import com.ourmenu.backend.domain.user.application.EmailService;
 import com.ourmenu.backend.domain.user.application.UserService;
 import com.ourmenu.backend.domain.user.domain.CustomUserDetails;
 import com.ourmenu.backend.domain.user.dto.*;
 import com.ourmenu.backend.global.response.ApiResponse;
 import com.ourmenu.backend.global.response.util.ApiUtil;
-import com.ourmenu.backend.global.util.JwtTokenProvider;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -17,9 +15,7 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/api/users")
 public class UserController {
 
-    private final JwtTokenProvider jwtTokenProvider;
     private final UserService userService;
-    private final EmailService emailService;
 
     /**
      * @apiNote 회원가입 API
@@ -42,18 +38,6 @@ public class UserController {
     private ApiResponse<SignInResponse> signIn(@RequestBody SignInRequest request, HttpServletResponse response){
         SignInResponse signInResponse = userService.signIn(request, response);
         return ApiUtil.success(signInResponse);
-    }
-
-    @PostMapping("/emails")
-    private ApiResponse<EmailResponse> sendConfirmCode(@RequestBody EmailRequest request){
-        EmailResponse response = emailService.sendCodeToEmail(request);
-        return ApiUtil.success(response);
-    }
-
-    @PostMapping("/emails/confirm-code")
-    private ApiResponse<String> verifyEmail(@RequestBody VerifyEmailRequest request){
-        String response = emailService.verifyConfirmCode(request);
-        return ApiUtil.success(response);
     }
 
     @PatchMapping("/password")
